Wanikani items refactoring
