+++ src/wanikani/components/WanikaniItemsControlPanel.jsx
@@ -1,169 +1,7 @@
 import {Checkbox, Paper, ToggleButton, ToggleButtonGroup} from "@mui/material";
 import * as React from "react";
-import {getColorByWanikaniSrsStage, getColorByWanikaniSubjectType} from "../service/WanikaniStyleUtil.js";
-import {useMemo} from "react";
-import kanji from "kanji";
-import {kanjiFrequencyLookupMap} from "../../util/KanjiDataUtil.js";
-
-export const groupByOptions = {
-    none: {
-        key: 'none',
-        displayText: 'None',
-        group: (subjects) => [
-            {
-                title: 'All Items',
-                subjects: subjects,
-            }
-        ],
-    },
-    level: {
-        key: 'level',
-        displayText: 'Level',
-        group: (subjects) => {
-            let levelsData = {};
-
-            for (let subject of subjects) {
-                if (!levelsData[subject.level]) {
-                    levelsData[subject.level] = [];
-                }
-                levelsData[subject.level].push(subject);
-            }
-
-            function getWanikaniLevels() {
-                return Array.from({length: 60}, (_, i) => i + 1);
-            }
-
-            return getWanikaniLevels()
-                .map(level => ({
-                    title: 'Level ' + level,
-                    subjects: levelsData[level] ?? []
-                }))
-                .filter(group => group.subjects.length > 0);
-        },
-    },
-    srsStage: {
-        key: 'srsStage',
-        displayText: 'SRS Stage',
-        group: (subjects) => {
-            let stageMap = {};
-
-            for (let subject of subjects) {
-                let srsStage = subject['srs_stage'] != 0 && !subject['srs_stage'] ? 'locked' : subject['srs_stage'];
-                if (!stageMap[srsStage]) {
-                    stageMap[srsStage] = [];
-                }
-                stageMap[srsStage].push(subject);
-            }
-
-            return ['Unlocked', 'Apprentice 1', 'Apprentice 2', 'Apprentice 3', 'Apprentice 4',
-                'Guru 1', 'Guru 2', 'Master', 'Enlightened', 'Burned', 'Locked']
-                .map((stage, index) => ({
-                    title: stage,
-                    subjects: stageMap[stage === 'Locked' ? 'locked' : index] ?? []
-                }))
-                .filter(group => group.subjects.length > 0);
-        },
-    },
-    jtpt: {
-        key: 'jlpt',
-        displayText: 'JLPT',
-        group: (subjects) => {
-
-            function toMap(array) {
-                let map = {};
-                for (const value of array) {
-                    map[value] = true;
-                }
-                return map;
-            }
-
-            const jtlp = [
-                toMap(kanji.jlpt.n5),
-                toMap(kanji.jlpt.n4),
-                toMap(kanji.jlpt.n3),
-                toMap(kanji.jlpt.n2),
-                toMap(kanji.jlpt.n1),
-            ];
-
-            let map = {};
-
-            for (let subject of subjects) {
-                let idx = jtlp.findIndex(lvl => lvl[subject['slug']]);
-                if (!map[idx]) {
-                    map[idx] = [];
-                }
-                map[idx].push(subject);
-            }
-
-            return ['N5', 'N4', 'N3', 'N2', 'N1', 'Non-JLPT']
-                .map((level, index) => ({
-                    title: level,
-                    subjects: map[level === 'Non-JLPT' ? -1 : index] ?? []
-                }))
-                .filter(group => group.subjects.length > 0);
-        },
-    },
-};
-
-export const sortByOptions = {
-    none: {
-        key: 'none',
-        displayText: 'None',
-        sort: (subjects) => subjects
-    },
-    itemName: {
-        key: 'itemName',
-        displayText: 'Item Name',
-        sort: (subjects) => subjects.sort((a, b) => a.slug.toLowerCase().localeCompare(b.slug.toLowerCase()))
-    },
-    level: {
-        key: 'level',
-        displayText: 'Level',
-        sort: (subjects) => subjects.sort((a, b) => a.level - b.level)
-    },
-    srsStage: {
-        key: 'srsStage',
-        displayText: 'SRS Stage',
-        sort: (subjects) => subjects.sort((a, b) => (a['srs_stage'] ?? 100) - (b['srs_stage'] ?? 100))
-    },
-    itemType: {
-        key: 'itemType',
-        displayText: 'Item Type',
-        sort: (subjects) => {
-            const typeOrder = {
-                radical: 1,
-                kanji: 2,
-                vocabulary: 3
-            };
-            return subjects.sort((a, b) => typeOrder[a.subjectType] - typeOrder[b.subjectType]);
-        }
-    },
-    frequency: {
-        key: 'frequency',
-        displayText: 'Frequency',
-        sort: (subjects) => {
-
-            function getFrequency(subject) {
-                return kanjiFrequencyLookupMap[subject.slug];
-            }
-
-            return subjects.sort((a, b) => getFrequency(a) - getFrequency(b));
-        }
-    },
-};
-
-export const colorByOptions = {
-    srsStage: {
-        key: 'srsStage',
-        displayText: 'SRS Stage',
-        color: (subject) => getColorByWanikaniSrsStage(subject['srs_stage'])
-    },
-    itemType: {
-        key: 'itemType',
-        displayText: 'Item Type',
-        color: (subject) => getColorByWanikaniSubjectType(subject.subjectType)
-    },
-};
+import {useMemo, useState} from "react";
+import {groupByOptions, sortByOptions, colorByOptions} from "../service/WanikaniDataUtil.js";
 
 const styles = {
     groupingPaper: {
@@ -181,26 +19,6 @@ const styles = {
         minWidth: '85px'
     }
 };
-
-function GroupByToggle({options, groupBy, setGroupBy, disableOptions}) {
-    return (
-        <ToggleButtonGroup
-            value={groupBy.key}
-            size={'small'}
-            exclusive
-            onChange={e => setGroupBy(options.find(o => o.key === e.target.value))}
-        >
-            {options.map((option) => (
-                <ToggleButton key={option.key}
-                              value={option.key}
-                              disabled={disableOptions?.map(o => o?.key).includes(option.key)}
-                >
-                    {option.displayText}
-                </ToggleButton>
-            ))}
-        </ToggleButtonGroup>
-    );
-}
 
 const groupByOptionsList = [
     groupByOptions.none,
@@ -221,6 +39,26 @@ const colorByOptionsList = [
     colorByOptions.itemType,
     colorByOptions.srsStage,
 ];
+
+function GroupByToggle({options, groupBy, setGroupBy, disableOptions}) {
+    return (
+        <ToggleButtonGroup
+            value={groupBy.key}
+            size={'small'}
+            exclusive
+            onChange={e => setGroupBy(options.find(o => o.key === e.target.value))}
+        >
+            {options.map((option) => (
+                <ToggleButton key={option.key}
+                              value={option.key}
+                              disabled={disableOptions?.map(o => o?.key).includes(option.key)}
+                >
+                    {option.displayText}
+                </ToggleButton>
+            ))}
+        </ToggleButtonGroup>
+    );
+}
 
 function ControlContainer({children}) {
     return (
@@ -288,21 +126,44 @@ function CheckboxControl({title, subtitle, value, setValue, options}) {
     );
 }
 
-function WanikaniItemsControlPanel(props) {
+export function useWanikaniItemControls() {
+    const [control, setControl] = useState({
+        primaryGroupBy: groupByOptions.level,
+        secondaryGroupBy: groupByOptions.none,
+        sortBy: sortByOptions.itemName,
+        colorBy: colorByOptions.itemType,
+        typesToShow: ['kanji']
+    });
+
+    return [
+        control,
+        {
+            control: setControl,
+            primaryGroupBy: (groupBy) => setControl(prev => ({...prev, primaryGroupBy: groupBy})),
+            secondaryGroupBy: (groupBy) => setControl(prev => ({...prev, secondaryGroupBy: groupBy})),
+            sortBy: (sortBy) => setControl(prev => ({...prev, sortBy: sortBy})),
+            colorBy: (colorBy) => setControl(prev => ({...prev, colorBy: colorBy})),
+            typesToShow: (typesToShow) => setControl(prev => ({...prev, typesToShow: typesToShow})),
+        }
+    ];
+}
+
+function WanikaniItemsControlPanel({control, set}) {
 
     const {
-        primaryGroupBy, setPrimaryGroupBy,
-        secondaryGroupBy, setSecondaryGroupBy,
-        sortBy, setSortBy,
-        colorBy, setColorBy,
-        typesToShow, setTypesToShow
-    } = props
+        primaryGroupBy,
+        secondaryGroupBy,
+        sortBy,
+        colorBy,
+        typesToShow
+    } = control;
 
     function onPrimaryGroupByChange(value) {
+        let controlChanges = {primaryGroupBy: value};
         if (value.key === groupByOptions.none.key || value.key === secondaryGroupBy.key) {
-            setSecondaryGroupBy(groupByOptions.none)
+            controlChanges.secondaryGroupBy = groupByOptions.none;
         }
-        setPrimaryGroupBy(value);
+        set.control(prev => ({...prev, ...controlChanges}))
     }
 
     const secondaryGroupDisabled = useMemo(() => {
@@ -330,7 +191,7 @@ function WanikaniItemsControlPanel(props) {
                     <div style={styles.optionLabel}>Secondary</div>
                     <GroupByToggle options={groupByOptionsList}
                                    groupBy={secondaryGroupBy}
-                                   setGroupBy={setSecondaryGroupBy}
+                                   setGroupBy={set.secondaryGroupBy}
                                    disableOptions={secondaryGroupDisabled}
                     />
                 </div>
@@ -339,20 +200,20 @@ function WanikaniItemsControlPanel(props) {
 
             <SegmentControl title={'Sort By'}
                             value={sortBy}
-                            setValue={setSortBy}
+                            setValue={set.sortBy}
                             options={sortByOptionsList}
             />
 
             <SegmentControl title={'Color By'}
                             value={colorBy}
-                            setValue={setColorBy}
+                            setValue={set.colorBy}
                             options={colorByOptionsList}
             />
 
             <CheckboxControl title={'Display'}
                              subtitle={'Types'}
                              value={typesToShow}
-                             setValue={setTypesToShow}
+                             setValue={set.typesToShow}
                              options={['Radical', 'Kanji', 'Vocabulary']}
             />
